SAAVI-APP/backend/requirements.txt
pydantic
python-dotenv
passlib[bcrypt]
requests
websocket-client
websockets
mpremote
